Relicense the project under MIT
Well, the main reason I used a (much) stronger MPL-2.0 is because its nice disclose source property. I was trying to emphasise that this project is a toy and the codebase is not meant for production use, so anyone can see how badly it is written if someone decided to use the code in a serious project. But it turns out to be unnecessary.